Legger InngangTilSykefraværsstatistikk som egen komponent for å gjøre ting cleanere
