+++ client/src/Infographic/datauthenting.ts
@@ -9,19 +9,19 @@ export function hentUtInfographicData(
   data: AggregertStatistikkDto
 ): InfographicData {
   const fraværsprosentNorge: StatistikkDto | undefined =
-    data.prosentSiste4Kvartaler.find(
+    data.prosentSiste4Kvartaler?.find(
       (s) => s.statistikkategori == Statistikkategori.LAND
     );
-  const fraværsprosentNæring = data.prosentSiste4Kvartaler.find(
+  const fraværsprosentNæring = data.prosentSiste4Kvartaler?.find(
     (s) => s.statistikkategori == Statistikkategori.NÆRING
   );
-  const fraværsprosentBransje = data.prosentSiste4Kvartaler.find(
+  const fraværsprosentBransje = data.prosentSiste4Kvartaler?.find(
     (s) => s.statistikkategori == Statistikkategori.BRANSJE
   );
-  const trendBransje = data.trend.find(
+  const trendBransje = data.trend?.find(
     (s) => s.statistikkategori == Statistikkategori.BRANSJE
   );
-  const trendNæring = data.trend.find(
+  const trendNæring = data.trend?.find(
     (s) => s.statistikkategori == Statistikkategori.NÆRING
   );
   return {
