+++ client/src/hooks/useMobileVersion.tsx
@@ -1,0 +1,8 @@
+import { useWindowSize } from "./useWindowSize";
+import styles from "../Infographic/Infographic.module.scss";
+
+export const useMobileVersion = () => {
+  const windowSize = useWindowSize();
+  const screenSmAsNumeric = parseInt(styles.screenSm.replace(/\D/g, ""));
+  return windowSize.width === undefined || windowSize.width < screenSmAsNumeric;
+};
