+++ client/src/integrasjoner/api-utils.ts
@@ -1,4 +1,6 @@
 import { RestRessurs, RestStatus, Årsak } from "./rest-status";
+import { Size } from "../hooks/useWindowSize";
+import styles from "../Infographic/Infographic.module.scss";
 
 export const getRestStatus = (responseStatus: number): RestStatus => {
   switch (responseStatus) {
@@ -31,7 +33,7 @@ export const fetchMedFeilhåndtering = async <T>(
     };
   }
   if (restStatus === RestStatus.Feil) {
-     //Sentry.captureException(new Error('Status ' + response.status + ' ved kall til ' + url));
+    //Sentry.captureException(new Error('Status ' + response.status + ' ved kall til ' + url));
 
     try {
       const body = await response.json();
